requirements.txt
streamlit>=1.28.0
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0
pillow>=9.0.0
requests>=2.28.0
opencv-python-headless>=4.5.0
numpy>=1.21.0,<2.0.0
ultralytics>=8.0.0
